Tweaks to DynScope generation.
* Limit specialized subclasses to 50 wide.
* Move to a generator class.
* Cache construction logic to StaticScope.
* Cache constructor in StaticScope.
* Reduce duplication in code gen.
* Move offset error to a shared method in generated classes.
* Use a single OneShotClassLoader since we never dereference.
* Misc perf, style tweaks.
