+++ core/src/main/java/org/jruby/runtime/DynamicScope.java
@@ -39,6 +39,7 @@ import org.jruby.runtime.builtin.IRubyObject;
 
 import static org.jruby.util.CodegenUtils.*;
 
+import org.jruby.runtime.scope.ManyVarsDynamicScope;
 import org.jruby.util.OneShotClassLoader;
 import org.jruby.util.collections.NonBlockingHashMapLong;
 
@@ -67,299 +68,7 @@ public abstract class DynamicScope implements Cloneable {
     }
 
     public static DynamicScope newDynamicScope(StaticScope staticScope, DynamicScope parent) {
-        return construct(staticScope, parent);
-    }
-
-    private static final NonBlockingHashMapLong<MethodHandle> prototypes = new NonBlockingHashMapLong<>();
-
-    public static MethodHandle protoClassFromProps(int size) {
-        return prototypes.get(size);
-    }
-
-    public static DynamicScope construct(StaticScope staticScope, DynamicScope parent) {
-        MethodHandle tupleType = generate(staticScope.getNumberOfVariables());
-        try {
-            return (DynamicScope) tupleType.invokeExact(staticScope, parent);
-        } catch (Throwable e) {
-            Helpers.throwException(e);
-            return null; // not reached
-        }
-    }
-
-    public static final List<String> SPECIALIZED_GETS = Collections.unmodifiableList(Arrays.asList(
-            "getValueZeroDepthZero",
-            "getValueOneDepthZero",
-            "getValueTwoDepthZero",
-            "getValueThreeDepthZero",
-            "getValueFourDepthZero",
-            "getValueFiveDepthZero",
-            "getValueSixDepthZero",
-            "getValueSevenDepthZero",
-            "getValueEightDepthZero",
-            "getValueNineDepthZero"
-    ));
-
-    public static final List<String> SPECIALIZED_SETS = Collections.unmodifiableList(Arrays.asList(
-            "setValueZeroDepthZeroVoid",
-            "setValueOneDepthZeroVoid",
-            "setValueTwoDepthZeroVoid",
-            "setValueThreeDepthZeroVoid",
-            "setValueFourDepthZeroVoid",
-            "setValueFiveDepthZeroVoid",
-            "setValueSixDepthZeroVoid",
-            "setValueSevenDepthZeroVoid",
-            "setValueEightDepthZeroVoid",
-            "setValueNineDepthZeroVoid"
-    ));
-
-    public static MethodHandle generate(final int size) {
-        MethodHandle h = protoClassFromProps(size);
-
-        if (h != null) return h;
-
-        Class p = null;
-        final String clsPath = "org/jruby/runtime/scopes/DynamicScope" + size;
-        final String clsName = clsPath.replaceAll("/", ".");
-
-        try {
-            // create a new one
-            final String[] newFields = varList(size);
-
-            final String baseName = p(DynamicScope.class);
-
-            JiteClass jiteClass = new JiteClass(clsPath, baseName, new String[0]) {{
-                // parent class constructor
-                defineMethod("<init>", ACC_PUBLIC, sig(void.class, StaticScope.class, DynamicScope.class), new CodeBlock() {{
-                    aload(0);
-                    aload(1);
-                    aload(2);
-                    invokespecial(baseName, "<init>", sig(void.class, StaticScope.class, DynamicScope.class));
-                    voidreturn();
-                }});
-
-                // required overrides
-                defineMethod("getValue", ACC_PUBLIC, sig(IRubyObject.class, int.class, int.class), new CodeBlock() {{
-                    LabelNode parentCall = new LabelNode(new Label());
-                    LabelNode defaultError = new LabelNode(new Label());
-                    LabelNode[] cases = new LabelNode[size];
-
-                    for (int i = 0; i < size; i++) {
-                        cases[i] = new LabelNode(new Label());
-                    }
-
-                    line(0);
-
-                    iload(2); // depth
-                    ifne(parentCall);
-
-                    if (size > 0) {
-                        iload(1);
-                        tableswitch(0, size - 1, defaultError, cases);
-                        for (int i = 0; i < size; i++) {
-                            label(cases[i]);
-                            aload(0);
-                            getfield(clsPath, newFields[i], ci(IRubyObject.class));
-                            areturn();
-                        }
-                        label(defaultError);
-                    }
-
-                    line(1);
-
-                    newobj(p(RuntimeException.class));
-                    dup();
-                    ldc(clsName + " only supports scopes with " + size + " variables");
-                    invokespecial(p(RuntimeException.class), "<init>", sig(void.class, String.class));
-                    athrow();
-                    label(parentCall);
-
-                    line(2);
-
-                    aload(0);
-                    getfield(baseName, "parent", ci(DynamicScope.class));
-                    iload(1);
-                    iload(2);
-                    pushInt(1);
-                    isub();
-                    invokevirtual(baseName, "getValue", sig(IRubyObject.class, int.class, int.class));
-                    areturn();
-                }});
-
-                defineMethod("setValueVoid", ACC_PUBLIC, sig(void.class, IRubyObject.class, int.class, int.class), new CodeBlock() {{
-                    LabelNode parentCall = new LabelNode(new Label());
-                    LabelNode defaultError = new LabelNode(new Label());
-                    LabelNode[] cases = new LabelNode[size];
-
-                    for (int i = 0; i < size; i++) {
-                        cases[i] = new LabelNode(new Label());
-                    }
-
-                    line(3);
-                    iload(3); // depth
-                    ifne(parentCall);
-                    if (size > 0) {
-                        iload(2);
-                        tableswitch(0, size - 1, defaultError, cases);
-                        for (int i = 0; i < size; i++) {
-                            label(cases[i]);
-                            aload(0);
-                            aload(1);
-                            putfield(clsPath, newFields[i], ci(IRubyObject.class));
-                            voidreturn();
-                        }
-                        label(defaultError);
-                    }
-                    line(4);
-                    newobj(p(RuntimeException.class));
-                    dup();
-                    ldc(clsName + " only supports scopes with " + size + " variables");
-                    invokespecial(p(RuntimeException.class), "<init>", sig(void.class, String.class));
-                    athrow();
-                    label(parentCall);
-                    line(5);
-                    aload(0);
-                    getfield(baseName, "parent", ci(DynamicScope.class));
-                    aload(1);
-                    iload(2);
-                    iload(3);
-                    pushInt(1);
-                    isub();
-                    invokevirtual(baseName, "setValueVoid", sig(void.class, IRubyObject.class, int.class, int.class));
-                    voidreturn();
-                }});
-
-                // optional overrides
-                defineMethod("getValueDepthZero", ACC_PUBLIC, sig(IRubyObject.class, int.class), new CodeBlock() {{
-                    LabelNode defaultError = new LabelNode(new Label());
-                    LabelNode[] cases = new LabelNode[size];
-
-                    for (int i = 0; i < size; i++) {
-                        cases[i] = new LabelNode(new Label());
-                    }
-
-                    line(6);
-                    if (size > 0) {
-                        iload(1);
-                        tableswitch(0, size - 1, defaultError, cases);
-                        for (int i = 0; i < size; i++) {
-                            label(cases[i]);
-                            aload(0);
-                            getfield(clsPath, newFields[i], ci(IRubyObject.class));
-                            areturn();
-                        }
-                        label(defaultError);
-                    }
-                    line(1);
-                    newobj(p(RuntimeException.class));
-                    dup();
-                    ldc(clsName + " only supports scopes with " + size + " variables");
-                    invokespecial(p(RuntimeException.class), "<init>", sig(void.class, String.class));
-                    athrow();
-                }});
-
-                defineMethod("setValueDepthZeroVoid", ACC_PUBLIC, sig(void.class, IRubyObject.class, int.class), new CodeBlock() {{
-                    LabelNode defaultError = new LabelNode(new Label());
-                    LabelNode[] cases = new LabelNode[size];
-
-                    line(6);
-
-                    for (int i = 0; i < size; i++) {
-                        cases[i] = new LabelNode(new Label());
-                    }
-                    if (size > 0) {
-                        iload(2);
-                        tableswitch(0, size - 1, defaultError, cases);
-                        for (int i = 0; i < size; i++) {
-                            label(cases[i]);
-                            aload(0);
-                            aload(1);
-                            putfield(clsPath, newFields[i], ci(IRubyObject.class));
-                            voidreturn();
-                        }
-                        label(defaultError);
-                    }
-                    line(1);
-                    newobj(p(RuntimeException.class));
-                    dup();
-                    ldc(clsName + " only supports scopes with " + size + " variables");
-                    invokespecial(p(RuntimeException.class), "<init>", sig(void.class, String.class));
-                    athrow();
-                }});
-
-                for (int i = 0; i < SPECIALIZED_GETS.size(); i++) {
-                    final int offset = i;
-
-                    defineMethod(SPECIALIZED_GETS.get(offset), ACC_PUBLIC, sig(IRubyObject.class), new CodeBlock() {{
-                        line(6);
-
-                        if (size <= offset) {
-                            newobj(p(RuntimeException.class));
-                            dup();
-                            ldc(clsName + " only supports scopes with " + size + " variables");
-                            invokespecial(p(RuntimeException.class), "<init>", sig(void.class, String.class));
-                            athrow();
-                        } else {
-                            aload(0);
-                            getfield(clsPath, newFields[offset], ci(IRubyObject.class));
-                            areturn();
-                        }
-                    }});
-                }
-
-
-                for (int i = 0; i < SPECIALIZED_SETS.size(); i++) {
-                    final int offset = i;
-
-                    defineMethod(SPECIALIZED_SETS.get(offset), ACC_PUBLIC, sig(void.class, IRubyObject.class), new CodeBlock() {{
-                        line(6);
-
-                        if (size <= offset) {
-                            newobj(p(RuntimeException.class));
-                            dup();
-                            ldc(clsName + " only supports scopes with " + size + " variables");
-                            invokespecial(p(RuntimeException.class), "<init>", sig(void.class, String.class));
-                            athrow();
-                        } else {
-                            aload(0);
-                            aload(1);
-                            putfield(clsPath, newFields[offset], ci(IRubyObject.class));
-                            voidreturn();
-                        }
-                    }});
-                }
-
-                // fields
-                for (String prop : newFields) {
-                    defineField(prop, ACC_PUBLIC, ci(IRubyObject.class), null);
-                }
-            }};
-
-            p = defineClass(jiteClass);
-
-            MethodHandle mh = MethodHandles.lookup().findConstructor(p, MethodType.methodType(void.class, StaticScope.class, DynamicScope.class));
-            mh = mh.asType(MethodType.methodType(DynamicScope.class, StaticScope.class, DynamicScope.class));
-            MethodHandle previousMH = prototypes.putIfAbsent(size, mh);
-            if (previousMH != null) mh = previousMH;
-
-            return mh;
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private static Class defineClass(JiteClass jiteClass) {
-        return new OneShotClassLoader(Ruby.getClassLoader())
-                .defineClass(jiteClass.getClassName().replaceAll("/", "."), jiteClass.toBytes(JDKVersion.V1_7));
-    }
-
-    public static String[] varList(int size) {
-        String[] vars = new String[size];
-
-        for (int i = 0; i < size; i++) {
-            vars[i] = "var" + i;
-        }
-
-        return vars;
+        return staticScope.construct(parent);
     }
 
     public static DynamicScope newDynamicScope(StaticScope staticScope, DynamicScope parent, EvalType evalType) {
